IMGCNV: Tuned clear color to better match EPD
The '8th' color of the EPD is the "clear" one. It is
used to clean the display but is a color darker than
white, close to white human skin. So it is good to
have in the palette. This change adjust it to better
match the display. The value before was a little to
blue.